+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## devel
 
+* add cluster alloctable memory and cpu to summary
+* add cluster memory and cpu capacity to summary
+* add cluster version to summary
+
 ## 0.2.1
 
 * add error logging for failed yaml parses